fixed a gcc warning

+++ uwadv/source/resource/cutsloader.cpp
@@ -124,6 +124,6 @@ void ua_cutscene::load(const char *filename)
 
    fclose(fd);
 
-   curframe = -1;
+   curframe = (unsigned int)-1;
    img.create(width,height,0,0);
 }
